+++ src/app/components/map/component.tsx
@@ -1,13 +1,13 @@
 'useClient';
 import React from 'react';
-import { MapContainer, Polyline, TileLayer } from 'react-leaflet';
+import { MapContainer, TileLayer } from 'react-leaflet';
 import 'leaflet/dist/leaflet.css';
 
 import { MapContextProvider } from '@contexts/map-context';
 import { MapEvents } from '@components/map/events';
 import { ContextMenu } from '@components/menu/context-menu';
 import Line from '@components/map/line';
-import NavigationMenu from '@components/menu/navigation-menu';
+import NavigationWrapper from '@components/menu/navigation-wrapper';
 
 const MapComponent = () => {
   return (
@@ -24,7 +24,7 @@ const MapComponent = () => {
           url='https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png'
         />
         <ContextMenu />
-        <NavigationMenu />
+        <NavigationWrapper />
         <Line />
         <MapEvents />
       </MapContextProvider>
