+++ src/app/components/menu/route-list.tsx
@@ -1,0 +1,7 @@
+import { Box } from '@mui/material';
+
+const RouteList = () => {
+  return <Box></Box>;
+};
+
+export default RouteList;
